Extended joystick data to 16bit, making room for extra buttons.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@40009 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/arch/gtk3/joy-unix-usb.c
@@ -311,13 +311,13 @@ void usb_joystick(void)
             if (it->type == ITEM_HAT) {
                 val -= it->min_val;
                 if (val >= 0 && val <= 7) {
-                    joystick_set_value_or(i + 1, hat_or[val]);
+                    joystick_set_value_or(i + 1, (uint16_t)hat_or[val]);
                 }
             } else {
                 if (val <= it->min_val) {
-                    joystick_set_value_or(i + 1, it->min_or);
+                    joystick_set_value_or(i + 1, (uint16_t)it->min_or);
                 } else if (val > it->max_val) {
-                    joystick_set_value_or(i + 1, it->max_or);
+                    joystick_set_value_or(i + 1, (uint16_t)it->max_or);
                 }
             }
         }
